refactor: remove constants to another file
